show-map now working
