+++ server/database-credentials.php
@@ -1,6 +1,6 @@
 <?php
-$servername="localhost";
-$username= "root";
-$password= "";
-$database= "";
-?>	
+$servername = "localhost";
+$username = "root";
+$password = "";
+$database = "wifi_database";
+
